refactor: resolve CR comments
