+++ package.json
@@ -16,8 +16,6 @@
     "@magiclabs/wagmi-connector": "^1.1.3",
     "@next/font": "13.1.6",
     "@oceanprotocol/lib": "3.1.5-next.0",
-    "@storybook/client-api": "^7.6.0",
-    "@storybook/react": "^7.6.0",
     "@tippyjs/react": "^4.2.6",
     "@types/node": "^18.15.5",
     "@types/numeral": "^2.0.2",
